requirements.txt
**************************** Image Denoising *****************************
torch = 1.5.1
scipy = 1.5.0
python = 3.7.11
numpy = 1.20.3

**************************** Image Inpainting *****************************
matplotlib=3.5.1
numpy=1.21.2
pillow=9.1.0
python=3.9.7
torch=1.10.2
scipy=1.8.0